add basic drinks table

+++ app/model.py
@@ -1,6 +1,9 @@
 from typing import Optional
+from datetime import datetime
+from uuid import uuid4
 
-from sqlalchemy import String, create_engine
+from sqlalchemy import ForeignKey, create_engine
+from sqlalchemy import types
 from sqlalchemy.orm import (
     DeclarativeBase,
     Mapped,
@@ -9,6 +12,8 @@ from sqlalchemy.orm import (
     scoped_session,
 )
 from typing_extensions import Self
+
+from app.config import is_debug_enabled
 
 
 engine = create_engine("sqlite:///water.sqlite")
@@ -30,7 +35,7 @@ class User(Base):
     __tablename__ = "user"
 
     id: Mapped[int] = mapped_column(primary_key=True)
-    name: Mapped[str] = mapped_column(String, nullable=False, unique=True)
+    name: Mapped[str] = mapped_column(types.String, nullable=False, unique=True)
     fullname: Mapped[Optional[str]]
     weight: Mapped[int]
     climate: Mapped[str]
@@ -46,6 +51,19 @@ class User(Base):
         return query.first()
 
 
+class Drinks(Base):
+    __tablename__ = "drinks"
+
+    id: Mapped[str] = mapped_column(
+        types.Text(length=36), primary_key=True, default=lambda: str(uuid4())
+    )
+    user_id: Mapped[int] = mapped_column(ForeignKey("user.id"))
+    timestamp: Mapped[datetime] = mapped_column(types.DateTime, nullable=False)
+    amount: Mapped[int] = mapped_column(types.Integer, nullable=False)
+
+
 def init_db():
-    Base.metadata.drop_all(engine)
+    """Initialize DB tables"""
+    if is_debug_enabled():
+        Base.metadata.drop_all(engine)
     Base.metadata.create_all(engine)
